# openapi.yaml
openapi: 3.1.0
info:
  title: Meraki License Server API
  version: 1.0.0
  description: |
    REST API for managing restaurant-CMS licenses.
    - **Public** endpoints (`/api/v1/`) — no auth, used by CMS instances
    - **Portal** endpoints (`/api/portal/`) — JWT Bearer (customer session)
    - **Admin** endpoints (`/api/admin/`) — JWT Bearer (admin session)
servers:
  - url: /api/v1
    description: Public API
  - url: /api/portal
    description: Customer Portal API
  - url: /api/admin
    description: Admin API
...
components:
  securitySchemes:
    portalAuth:
      type: http
      scheme: bearer
      description: Customer portal JWT (from POST /api/portal/login)
    adminAuth:
      type: http
      scheme: bearer
      description: Admin JWT (from POST /api/admin/login)
    licenseAuth:
      type: http
      scheme: bearer
      description: Signed license JWT (from POST /api/v1/validate)
...
paths:
...
  /admin/settings/rotate-keys:
    post:
      tags: [admin]
      summary: Rotate RSA signing key (SuperAdmin)
      description: Generates new RSA-2048 pair, retires old key. Existing tokens stay valid via kid lookup.
      security: [adminAuth: []]
      responses:
        '200':
          description: New key active
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  kid: { type: string }
  /admin/settings/sessions:
    get:
      tags: [admin]
      summary: List active admin sessions (SuperAdmin)
      security: [adminAuth: []]
      responses:
        '200': { description: Active session list }
...
  /admin/settings/webhook-dead-letters:
    get:
      tags: [admin]
      summary: List unresolved webhook dead letters (SuperAdmin)
      security: [adminAuth: []]
      responses:
        '200': { description: Dead letter list }
  /admin/settings/webhook-dead-letters/{id}/retry:
    post:
      tags: [admin]
      summary: Retry a failed webhook delivery (SuperAdmin)
      security: [adminAuth: []]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: Retry successful }
        '502': { description: Retry failed again }
...